redux and redesign start
