+++ src/client/index.css
@@ -49,9 +49,99 @@ h1 {
 /* ////////////////////////////////// */
 
 /* ////////////////////////////////// */
+/* Container for the image and fade effect */
+.menu-art-container {
+  position: relative;
+  width: 100%;
+  height: 80vh;
+  overflow: hidden;
+  margin-bottom: 2em;
+}
+
+/* .menu-art-container svg {
+  pointer-events: all;
+} */
+
+.menu-art-container polygon {
+  transition: fill 0.2s;
+  cursor: pointer;
+  pointer-events: all;
+}
+
+.menu-art-container polygon:hover {
+  fill-opacity: 0.8;
+}
+/* Fading overlay effect */
+
+.menu-art-container::before {
+  content: "";
+  position: absolute;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  background: linear-gradient(
+    to right,
+    rgba(255, 255, 255, 1) 0%,
+    rgba(255, 255, 255, 0) 3%,
+    rgba(255, 255, 255, 0) 97%,
+    rgba(255, 255, 255, 1) 100%
+  );
+  pointer-events: none; 
+}
+
+/* Adding the map and area to ensure interactions are preserved */
+#menu-art {
+  position: absolute;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  
+}
+
+/* ////////////////////////////////// */
+/* ////////////////////////////////// */
+/* ////////////////////////////////// */
+/* ////////////////////////////////// */
+/* ////////////////////////////////// */
+/* ////////////////////////////////// */
+/* ////////////////////////////////// */
+/* ////////////////////////////////// */
+
+/* .menu-art-container {
+  max-width: 2000px;
+  margin: 0 auto; 
+  overflow: hidden;
+}
+
+.menu-art-container img {
+  width: 100%;
+  max-width: 2000px;
+  height: auto;
+  object-fit: cover;
+}
+
+@media (max-width: 1200px) {
+  .menu-art-container img {
+    width: 1200px;
+  }
+} */
+
+
+/* ////////////////////////////////// */
+/* ////////////////////////////////// */
+/* ////////////////////////////////// */
+/* ////////////////////////////////// */
+/* ////////////////////////////////// */
+/* ////////////////////////////////// */
+/* ////////////////////////////////// */
+/* ////////////////////////////////// */
+
+
+
 #main-container {
   margin: 0 10vw;
-  /* border: 1px solid black; */
   display: flex;
   direction: column;
   justify-content: left;
@@ -62,6 +152,7 @@ h1 {
   display: flex;
   direction: column;
   justify-content: center;
+  flex-wrap: wrap;
 }
 
 .sub-header {
@@ -144,7 +235,6 @@ h1 {
   margin: 10px;
 }
 
-
 /* ------------------------------------ */
 
 .custom-carousel .slick-prev,
@@ -161,7 +251,6 @@ footer {
   width: 100%;
 } */
 
-
 @media screen and (max-width: 800px) {
   .big-art {
     height: 60vh;
@@ -177,10 +266,20 @@ footer {
     direction: column;
     justify-content: left;
   }
-  
 
   .sub-header {
     width: 100%;
   }
 
+  .menu-art-container {
+    height: 60vh;
+  }
+
+}
+
+/* iPhone12 */
+@media screen and (max-width: 400px) {
+  .menu-art-container {
+    height: 50vh;
+  }
 }
